requirements.txt
pyannote.audio
ipykernel
ipywidgets
simpleder